feat: base struct of packages organized
