+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/infra/configs/DbConfig.java
@@ -1,4 +1,4 @@
-package com.ciandt.summit.bootcamp2022.config;
+package com.ciandt.summit.bootcamp2022.infra.configs;
 
 import java.util.Properties;
 
@@ -7,10 +7,7 @@ import javax.sql.DataSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.context.annotation.PropertySource;
 import org.springframework.core.env.Environment;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.jdbc.datasource.DriverManagerDataSource;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
