+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/infra/configs/SqlDialect.java
@@ -1,4 +1,4 @@
-package com.ciandt.summit.bootcamp2022.config;
+package com.ciandt.summit.bootcamp2022.infra.configs;
 
 import java.sql.Types;
 
@@ -6,7 +6,6 @@ import org.hibernate.dialect.Dialect;
 import org.hibernate.dialect.function.StandardSQLFunction;
 import org.hibernate.dialect.function.SQLFunctionTemplate;
 import org.hibernate.dialect.function.VarArgsSQLFunction;
-import org.hibernate.Hibernate;
 import org.hibernate.type.StringType;
 
 public class SqlDialect extends Dialect {
